requirements.txt
Django==3.1.6
django-allauth==0.44.0
django-environ==0.4.5
django-extensions==3.1.1
psycopg2-binary==2.8.6
ipdb==0.13.8
ipython==7.24.0
django-model-utils==4.1.1
marshmallow==3.12.1 
inertia-django==0.3.2
django-js-routes==0.1.3
PyYAML==5.4.1